+++ lib/wright/dsl.rb
@@ -32,17 +32,17 @@ module Wright
     # Creates a resource method in the DSL module. Uses the
     # snake-cased class name as method name.
     #
-    # Typically resource_class is a subclass of {Resource}. It is
-    # initialized with the resource's name as an argument.
+    # Typically +resource_class+ is a subclass of {Resource}. It is
+    # initialized with the resource's name and the attribute hash as
+    # arguments.
     #
-    # @param resource_class the resource class
-    #
+    # @param resource_class [Class] the resource class
     # @return [void]
     def self.register_resource(resource_class)
       method_name = Util.class_to_resource_name(resource_class)
       this_module = self
-      define_method(method_name) do |name, &block|
-        this_module.yield_resource(resource_class, name, &block)
+      define_method(method_name) do |name, args = {}, &block|
+        this_module.yield_resource(resource_class, name, args, &block)
       end
     end
 
@@ -52,10 +52,14 @@ module Wright
     # Implicitly invoking a block from within another block does not
     # work: http://blog.sidu.in/2007/11/ruby-blocks-gotchas.html
     #
+    # @param resource_class [Class] the resource class
+    # @param name [String] the name of the resource object
+    # @param args [Hash<Symbol, Object] the attribute hash of the resource
+    #
     # @yield [Resource] the resource
     # @return [void]
-    def self.yield_resource(resource_class, name)
-      r = resource_class.new(name)
+    def self.yield_resource(resource_class, name, args)
+      r = resource_class.new(name, args)
       yield(r) if block_given?
       r.run_action if r.respond_to?(:run_action)
       r