+++ lib/wright/util/file_owner.rb
@@ -9,6 +9,11 @@ module Wright
       # @return [String, Integer] the group's name or gid
       attr_accessor :group
 
+      def initialize(user_and_group = nil, group = nil)
+        self.user_and_group = user_and_group
+        self.group = group if group
+      end
+
       # Sets user and group simultaneously.
       #
       # @param user_and_group [String, Integer] a user in +user:group+