+++ spec/resource/directory_spec.rb
@@ -9,6 +9,21 @@ describe Wright::Resource::Directory do
 
   after(:each) { FakeFS::FileSystem.clear }
 
+  describe '#initialize' do
+    it 'should accept attributes via an argument hash' do
+      resource = Wright::Resource::Directory.new(@dirname,
+                                                 mode: 'sample_mode',
+                                                 action: 'sample_action',
+                                                 owner: 'sample_owner',
+                                                 group: 'sample_group')
+      resource.name.must_equal @dirname
+      resource.mode.must_equal 'sample_mode'
+      resource.action.must_equal 'sample_action'
+      resource.owner.must_equal 'sample_owner'
+      resource.group.must_equal 'sample_group'
+    end
+  end
+
   describe '#create' do
     it 'should create directories' do
       FakeFS do