+++ spec/resource/group_spec.rb
@@ -9,6 +9,21 @@ describe Wright::Resource::Group do
     @group.instance_variable_set(:@provider, @provider)
   end
 
+  describe '#initialize' do
+    it 'should accept attributes via an argument hash' do
+      resource = Wright::Resource::Group.new('sample_name',
+                                             members: 'sample_members',
+                                             gid: 'sample_gid',
+                                             system: 'sample_system',
+                                             action: 'sample_action')
+      resource.name.must_equal 'sample_name'
+      resource.members.must_equal 'sample_members'
+      resource.gid.must_equal 'sample_gid'
+      resource.system.must_equal 'sample_system'
+      resource.action.must_equal 'sample_action'
+    end
+  end
+
   describe '#create' do
     it 'should ask the provider to create the group' do
       @provider.expect(:create, nil)