+++ NEWS.md
@@ -1,4 +1,7 @@
 # wright NEWS
+## unreleased
+- Add attribute hash support to DSL and resources
+
 ## 0.4.2 (2015-06-15)
 - Don't colorize log messages if Config[:log][:colorize] is nil (#13)